# Conflicts:
#	NandradSolver/src/NM_HydraulicNetworkFlowElements.cpp
#	NandradSolver/src/NM_ThermalNetworkFlowElements.cpp
#	SIM-VICUS/projects/Qt/SIM-VICUS.pro
#	SIM-VICUS/src/SVDatabase.cpp
#	SIM-VICUS/src/SVMainWindow.cpp
#	SIM-VICUS/src/SVMainWindow.h
#	SIM-VICUS/src/SVMainWindow.ui
#	SIM-VICUS/src/SVPropModeSelectionWidget.h
#	SIM-VICUS/src/SVSimulationStartNandrad.cpp
#	SIM-VICUS/src/SVSimulationStartNandrad.h
#	SIM-VICUS/src/SVSimulationStartNandrad.ui
#	SIM-VICUS/src/core3D/Vic3DScene.cpp
#	externals/Vicus/src/VICUS_KeywordList.cpp
#	externals/Vicus/src/VICUS_Project.h
#	externals/Vicus/src/VICUS_SubNetwork.h
#	externals/Vicus/src/ncg/ncg_VICUS_SubNetwork.cpp
#	externals/Vicus/srcTranslations/VICUS_KeywordListQt.cpp

+++ externals/IBK/src/IBK_InputOutput.cpp
@@ -227,7 +227,7 @@ void read_string_list_binary( std::istream &in,
 				// if line doesn't contain end marker append to list
 				// otherwise break execution now
 				// append string to the string list
-				unsigned int comparisonLimiter = lastLineTokens.size();
+				unsigned int comparisonLimiter = (unsigned int)lastLineTokens.size();
 				for ( unsigned int r = 0; r < comparisonLimiter; ++r ){
 
 					std::string::size_type pos = line.find( lastLineTokens[r] );
@@ -365,7 +365,7 @@ void read_string_vector_binary( std::istream &in,
 				// if line doesn't contain end marker append to list
 				// otherwise break execution now
 				// append string to the string list
-				unsigned int comparisonLimiter = lastLineTokens.size();
+				unsigned int comparisonLimiter = (unsigned int)lastLineTokens.size();
 				for ( unsigned int r = 0; r < comparisonLimiter; ++r ){
 
 					std::string::size_type pos = line.find( lastLineTokens[r] );
